sentiment-analysis/requirements.txt
quixstreams<2.7
transformers[torch]
python-dotenv